requirements.txt
# Core Trading and Financial APIs
alpaca-trade-api>=3.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Data Analysis and Processing
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b0
talib>=0.4.19
scipy>=1.11.0

# Web Scraping and HTML Processing
beautifulsoup4>=4.12.0

# Google Sheets Integration
gspread>=5.0.0
oauth2client>=4.1.3

# AI and Natural Language Processing
openai>=1.0.0

# Environment and Configuration Management
python-dotenv>=1.0.0

# Timezone Handling
pytz>=2023.3

# Network and HTTP
urllib3>=2.0.0

# System and Process Monitoring
psutil>=5.9.0

# Data Visualization
matplotlib>=3.7.0

# Concurrency and Threading
# (Built-in modules: asyncio, threading, concurrent.futures)

# Development and Testing Dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Type Checking (optional)
mypy>=1.5.0

# Code Quality (optional)
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0